% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_convert.R
\name{convert_exercise_to_nb}
\alias{convert_exercise_to_nb}
\title{Convert Exercise To Notebook}
\usage{
convert_exercise_to_nb(ps_ex_name, pb_force = FALSE)
}
\arguments{
\item{ps_ex_name}{exercise name}

\item{pb_force}{boolean flag to force overwrite the generated notebook}
}
\description{
Wrapper function to convert a RMarkdown exercise source file to a notebook
}
\details{
The wrapper assumes for convenience directory settings that are used in the
course project directory
}
\examples{
\dontrun{
devtools::load_all()
convert_exercise_to_nb(ps_ex_name = 'asm_ex02')
}

}
