% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_create.R
\name{create_exercise}
\alias{create_exercise}
\title{Generic Function For Creating New Exercise}
\usage{
create_exercise(
  ps_ex_name,
  ps_course_name,
  pn_exercise_count,
  ps_author,
  pn_nr_problem,
  pd_creation_date = format(Sys.Date(), "\%Y-\%m-\%d"),
  pb_edit = FALSE
)
}
\arguments{
\item{ps_ex_name}{name of the exercise}

\item{ps_course_name}{name of the course for which the exercise is created}

\item{pn_exercise_count}{the number of the exercise in the course}

\item{ps_author}{author of the exercise}

\item{pn_nr_problem}{the number of problems in the exercise}

\item{pd_creation_date}{date of creation of the exercise}

\item{pb_edit}{should created file be directly edited}
}
\description{
This is a convenience function for collecting the input for calling the
function \code{rteachtools::create_exercise()}.
}
\examples{
\dontrun{
create_exercise(ps_ex_name        = "gel_ex01",
                ps_course_name    = "Genetic Evaluation",
                pn_exercise_count = 1,
                ps_author         = "Peter von Rohr",
                pn_nr_problem     = 2)

}
}
