% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_create.R
\name{create_exercise_gel}
\alias{create_exercise_gel}
\title{Wrapper for Genetic Evaluation Exercises}
\usage{
create_exercise_gel(
  ps_ex_name,
  pn_exercise_count,
  ps_author,
  pn_nr_problem,
  pd_creation_date = format(Sys.Date(), "\%Y-\%m-\%d"),
  pb_edit = FALSE
)
}
\arguments{
\item{ps_ex_name}{name of the exercise}

\item{pn_exercise_count}{the number of the exercise in the course}

\item{ps_author}{autor of the exercise}

\item{pn_nr_problem}{number of problems in the exercise}

\item{pd_creation_date}{createion data of the exercise}

\item{pb_edit}{should created file be directly edited}
}
\description{
Wrapper for Genetic Evaluation Exercises
}
\examples{
\dontrun{
create_exercise_gel(ps_ex_name        = "gel_ex01",
                    pn_exercise_count = 1,
                    ps_author         = "Peter von Rohr",
                    pn_nr_problem     = 2)
}
}
