% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_deploy.R
\name{deploy_exercise}
\alias{deploy_exercise}
\title{Exercise Deployment}
\usage{
deploy_exercise(ps_ex_name)
}
\arguments{
\item{ps_ex_name}{exercicse name}
}
\description{
Wrapper for deployment of a given exercise in a course. The deployment
consists of producing the pdf-output of the exercise without solutions and
the pdf-document with the solutions.
}
\details{
This wrapper assumes that the source Rmd file containing the exercise problems
and the solutions is in a subdirectory called 'ex' under the course project directory.
}
\examples{
\dontrun{
devtools::load_all()
deploy_exercise(ps_ex_name = 'asm_ex02')
}

}
