% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ex_deploy.R
\name{deploy_uni_src}
\alias{deploy_uni_src}
\title{Deployment Based on Unified Source File}
\usage{
deploy_uni_src(ps_ex_name)
}
\arguments{
\item{ps_ex_name}{name of exercise}
}
\description{
The unified source file is modified by cutting out selected
sections which characteristic for certain output instance.
}
\examples{
\dontrun{
devtools::load_all();deploy_uni_src(ps_ex_name = 'asm_ex06')
}
}
